feat(M10): lib split + smoke test + clippy cleanup
src/lib.rs:
- Pure engine modules (voxel, lighting, mesher, physics, jobs,
  persistence, worldgen) exposed as a library so integration tests
  can drive them headlessly
- Binary-only modules (app, ecs, render) stay in main.rs to keep the
  library's dependency surface free of winit/wgpu
- main.rs uses 'pub use oxium::...' so existing 'crate::voxel::...'
  paths in the bin's submodules continue to resolve

tests/smoke.rs:
- generate_light_mesh_round_trip: 5x3x5 chunks generated + lit, then
  decompressed; asserts every light value is in 0..=15
- greedy_mesh_a_generated_chunk: meshes a real worldgen chunk;
  asserts non-empty + index count divisible by 3

Polish per the plan's self-review:
- Removed MeshHandle + ChunkMeta::mesh_handles (renderer owns the
  actual GPU handles per the v0 design)
- ChunkState now derives Default (cleaner than the manual impl)
- Jobs derives Default
- raycast.rs: collapsed if-let-and; removed dead initial assignment
- world_stream uses Entry API instead of contains_key + insert
- LOD downsample: dropped a redundant u32->u32 cast
- Documented #[allow(dead_code)] on the kept-alive Gpu fields + the
  spec-modelled-but-unused Camera::fov / PlayerInput::toggle_mode
- #[allow(clippy::too_many_arguments)] on the few inherently-wide
  helpers (world_stream, render_to_view, emit_greedy_quad)

Skipped per spec's 'open questions' note:
- wgpu_text HUD (font URL returns 404; spec marks this as 'corner text
  in v0; hotbar with M10 polish if time')
- cargo flamegraph profile (needs interactive sudo; can't run unattended)

Result: cargo test -> 50 unit + 2 smoke passing; cargo clippy -D
warnings -> clean.

+++ src/mesher/lod.rs
@@ -67,7 +67,7 @@ pub fn downsample(src: &DenseChunk, factor: u32) -> LodChunk {
                     .max_by_key(|(_, c)| *c)
                     .map(|(b, _)| b)
                     .unwrap_or(Block::Air);
-                let n = (factor * factor * factor) as u32;
+                let n = factor * factor * factor;
                 let sky = (sky_sum / n) as u8;
                 let blk = (blk_sum / n) as u8;
                 let idx = (x + y * dim + z * dim * dim) as usize;
